v1_legacy/requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
# langchain-ollama>=0.1.0  # Uncomment to use local Ollama
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
fpdf2>=2.7.0
python-docx>=0.8.11
python-pptx>=0.6.21
